+++ templates/config-page.php
@@ -50,19 +50,6 @@
                 </div>
             </div>
         </div>
-
-        <!--
-        <div class="small-12 medium-3 large-3 columns">
-          <p>Mostrar Logos</p>
-          <div class="switch small">
-            <input class="switch-input" type="checkbox" name="showlogo" id="showlogo" style="margin-top: 10px" <?php #echo ($showlogo === true) ? 'checked' : ''; ?>>
-            <label class="switch-paddle" for="showlogo">
-              <span class="switch-active" aria-hidden="true">Si</span>
-              <span class="switch-inactive" aria-hidden="true">No</span>
-            </label>
-          </div>
-        </div>
-        -->
     </div>
 
     <div class="row">
@@ -79,14 +66,6 @@
             <label for="privatekey">
                 Llave Privada
                 <input type="text" name="privatekey" id="privatekey" placeholder="sk_live_xxxxxxxxxxxxxxx" value="<?php echo $privatekey; ?>">
-            </label>
-        </div>
-    </div>
-
-    <div class="row">
-        <div class="large-12 columns">
-            <label for="webhook">
-                <input type="hidden" name="webhook" id="webhook" value="<?php echo $webhook; ?>">
             </label>
         </div>
     </div>
@@ -114,33 +93,6 @@
                 </select>
             </label>
             <input type="button" value="Deshabilitar" id="quitar_proveedor" class="button primary expanded">
-        </div>
-    </div>
-
-    <div class="row">
-        <div class="large-12 columns">
-            <label for="title">
-                Titulo
-                <input type="text" name="title" id="title" value="<?php echo $titulo ? $titulo : 'ComproPago (OXXO, 7Eleven, etc.)'; ?>">
-            </label>
-        </div>
-    </div>
-
-    <div class="row">
-        <div class="large-12 columns">
-            <label for="descripsion">
-                Descripcion
-                <input type="text" name="descripcion" id="descripsion" value="<?php echo $descripcion ? $descripcion : 'Con ComproPago, realiza tus pagos en Oxxo, 7Eleven y mas...' ?>">
-            </label>
-        </div>
-    </div>
-
-    <div class="row">
-        <div class="large-12 columns">
-            <label for="select-text">
-                Instrucciones
-                <input type="text" name="select-text" id="instrucciones" value="<?php echo $instrucciones ? $instrucciones : 'Selecciona el establecimiento donde deseas pagar'; ?>">
-            </label>
         </div>
     </div>
 
@@ -174,6 +126,8 @@
             <input type="button" class="button primary expanded" value="Guardar configuracion" id="save-config-compropago">
         </div>
     </div>
+
+    <input type="hidden" name="webhook" id="webhook" value="<?php echo $webhook; ?>">
 </section>
 
 <section class="cp-block-login" id="loadig">
